ekstedta/dominion: implement Adventurer
Side note:
It turns out there are a couple treasure cards in expansions which are
also actions--one of which has a different effect depending on whether
you play it during the action phase or the buy phase. So i can't just
automatically count/play treasure cards for you like i had planned.